renamed TaxCalculatorPetrol to EmissionTaxCalculator, extracted logic for Petrol and Alternative fuel, added decision to calculateTax() method
