+++ exercises/java/src/test/java/tax/TaxCalculatorAlternativeFuelTest.java
@@ -19,7 +19,7 @@ public class TaxCalculatorAlternativeFuelTest {
 
     @Before
     public void setUp() {
-        taxCalculator = new DummyTaxCalculator();
+        taxCalculator = new EmissionTaxCalculator();
     }
 
     @Test
