+++ exercises/java/src/test/java/tax/TaxCalculatorPetrolTest.java
@@ -18,7 +18,7 @@ public class TaxCalculatorPetrolTest {
 
     @Before
     public void setUp() {
-        taxCalculator = new TaxCalculatorPetrol();
+        taxCalculator = new EmissionTaxCalculator();
     }
 
     @Test
